requirements.txt
appdirs==1.4.4
asgiref==3.4.1
autoflake==1.4
black==21.7b0
click==8.0.1
Django==3.2.6
-e git+git@github.com:Kel0/django-parrallel-sessions.git@e732242819511765726607e5978503703d059fb6#egg=dps
flake8==3.9.2
gitdb==4.0.7
GitPython==3.1.18
invoke==1.6.0
isort==5.9.3
mccabe==0.6.1
myhello==0.1
mypy==0.910
mypy-extensions==0.4.3
pathspec==0.9.0
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
redis==3.5.3
regex==2021.8.3
smmap==4.0.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.0
